+++ pkg/ruler/rules/store.go
@@ -24,7 +24,10 @@ var (
 // To make sure that rules within each group are loaded, client must use LoadRuleGroups method.
 type RuleStore interface {
 	ListAllUsers(ctx context.Context) ([]string, error)
+
+	// ListAllRuleGroups returns all rule groups for all users.
 	ListAllRuleGroups(ctx context.Context) (map[string]RuleGroupList, error)
+
 	// ListRuleGroupsForUserAndNamespace returns all the active rule groups for a user from given namespace.
 	// If namespace is empty, groups from all namespaces are returned.
 	ListRuleGroupsForUserAndNamespace(ctx context.Context, userID string, namespace string) (RuleGroupList, error)
@@ -36,8 +39,16 @@ type RuleStore interface {
 
 	GetRuleGroup(ctx context.Context, userID, namespace, group string) (*RuleGroupDesc, error)
 	SetRuleGroup(ctx context.Context, userID, namespace string, group *RuleGroupDesc) error
+
+	// DeleteRuleGroup deletes single rule group.
 	DeleteRuleGroup(ctx context.Context, userID, namespace string, group string) error
+
+	// DeleteNamespace lists rule groups for given user and namespace, and deletes all rule groups.
+	// If namespace is empty, deletes all rule groups for user.
 	DeleteNamespace(ctx context.Context, userID, namespace string) error
+
+	// SupportsModifications returns true if this RuleStore supports operations that modify rule groups in the store.
+	SupportsModifications() bool
 }
 
 // RuleGroupList contains a set of rule groups
@@ -63,6 +74,10 @@ type ConfigRuleStore struct {
 	configClient  client.Client
 	since         userconfig.ID
 	ruleGroupList map[string]RuleGroupList
+}
+
+func (c *ConfigRuleStore) SupportsModifications() bool {
+	return false
 }
 
 // NewConfigRuleStore constructs a ConfigRuleStore